fix(web): modularization of components, mobilecard, desktopcard, style changes
