+++ web/src/pages/Dashboard/Courts/CourtCard/LockedStake.tsx
@@ -1,0 +1,27 @@
+import React from "react";
+import styled, { css } from "styled-components";
+import { landscapeStyle } from "styles/landscapeStyle";
+import { formatUnits } from "viem";
+
+const StyledLockedStakeLabel = styled.label`
+  display: flex;
+  justify-content: flex-end;
+  color: ${({ theme }) => theme.primaryText};
+
+  ${landscapeStyle(
+    () => css`
+      min-width: 107px;
+    `
+  )}
+`;
+
+interface ILockedStake {
+  lockedStake: bigint;
+}
+
+const LockedStake: React.FC<ILockedStake> = ({ lockedStake }) => {
+  const formattedLockedStake = formatUnits(lockedStake, 18);
+
+  return <StyledLockedStakeLabel>{`${formattedLockedStake} PNK`}</StyledLockedStakeLabel>;
+};
+export default LockedStake;
