+++ web/src/pages/Dashboard/Courts/Header/Stake.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import styled, { css } from "styled-components";
+import { landscapeStyle } from "styles/landscapeStyle";
+
+const StyledStakeLabel = styled.label`
+  display: flex;
+
+  ${landscapeStyle(() => css``)}
+`;
+
+const Stake: React.FC = () => {
+  return <StyledStakeLabel> Stake </StyledStakeLabel>;
+};
+export default Stake;
