+++ web/src/pages/Dashboard/Courts/index.tsx
@@ -25,7 +25,7 @@ const Divider = styled.hr`
 const CourtCardsContainer = styled.div`
   display: flex;
   flex-direction: column;
-  gap: 8px;
+  gap: 16px;
   z-index: 0;
 `;
